Refactored the inject function and fleshed out the module
Inject function now calls a hashRegion, which is used to hash an entity
based on its size. removeEntityFromHashmap was added with the idea that
it will be used in conjunction with the eventual collision detection.
The idea being if a collision is detected, the enteries in the hashmap
will become outdated, so will be removed from the map, then its response
will be processed, and that data will be inserted into the hashmap.